+++ compiler/src/iree/compiler/Codegen/Common/BUILD.bazel
@@ -160,6 +160,7 @@ iree_compiler_cc_library(
         "FlattenMemRefSubspanPass.cpp",
         "FoldAffineMinInDistributedLoops.cpp",
         "FoldTensorExtractOpPass.cpp",
+        "FoldTensorSubsetIntoVectorTransferOps.cpp",
         "ForOpCanonicalizationPass.cpp",
         "FuseTensorPadWithConsumer.cpp",
         "GenericVectorization.cpp",
