+++ compiler/src/iree/compiler/Codegen/SPIRV/SPIRVVectorize.cpp
@@ -406,7 +406,7 @@ public:
       RewritePatternSet patterns(context);
       vector::TransferReadOp::getCanonicalizationPatterns(patterns, context);
       vector::TransferWriteOp::getCanonicalizationPatterns(patterns, context);
-      tensor::populateFoldTensorSubsetIntoVectorTransferPatterns(patterns);
+      populateVectorTransferTensorSliceTransforms(patterns);
 
       if (failed(applyPatternsAndFoldGreedily(funcOp, std::move(patterns)))) {
         return signalPassFailure();
@@ -555,7 +555,7 @@ public:
 
       vector::TransferReadOp::getCanonicalizationPatterns(patterns, context);
       vector::TransferWriteOp::getCanonicalizationPatterns(patterns, context);
-      tensor::populateFoldTensorSubsetIntoVectorTransferPatterns(patterns);
+      populateVectorTransferTensorSliceTransforms(patterns);
 
       if (failed(applyPatternsAndFoldGreedily(funcOp, std::move(patterns)))) {
         return signalPassFailure();
@@ -645,7 +645,7 @@ public:
       vector::ExtractOp::getCanonicalizationPatterns(patterns, context);
       vector::TransferReadOp::getCanonicalizationPatterns(patterns, context);
       vector::TransferWriteOp::getCanonicalizationPatterns(patterns, context);
-      tensor::populateFoldTensorSubsetIntoVectorTransferPatterns(patterns);
+      populateVectorTransferTensorSliceTransforms(patterns);
       vector::ReductionOp::getCanonicalizationPatterns(patterns, context);
       scf::IfOp::getCanonicalizationPatterns(patterns, context);
       if (failed(applyPatternsAndFoldGreedily(funcOp, std::move(patterns)))) {
