+++ compiler/src/iree/compiler/Codegen/TransformStrategies/Common/Common.cpp
@@ -346,9 +346,8 @@ Value mlir::iree_compiler::buildBufferize(ImplicitLocOpBuilder &b,
   buildCanonicalizationAndEnablingTransforms(
       b, funcH, [](OpBuilder &b, Location loc) {
         b.create<transform::ApplyReassociativeReshapeFoldingPatternsOp>(loc);
-        b.create<
-            transform::ApplyFoldTensorSubsetOpsIntoVectorTransfersPatternsOp>(
-            loc);
+        b.create<IREE::transform_dialect::
+                     ApplyFoldTensorSliceIntoTransferPatternsOp>(loc);
       });
   b.create<IREEEliminateEmptyTensorsOp>(variantH);
   variantH = b.create<IREEBufferizeOp>(variantH, targetGpu);
